More detail on factory vs singleton
Hi there, I'm not at all sure how to contact you about this as there's no comments on your blog. I hope this is ok... 

Sorry to gravedig an 8 year old post here, but I am the sort of person you wrote these articles for - a new, self-taught "developer" - and came across your blog while trying to wrap my head around OOP, singletons etc...

I'm quite sure that this isn't actually the case, but your two examples for singleton and factory seem identical to me.

This is how it appears to me, so I'm hoping you can clarify here (or, better yet, in a blog post for others to more easily learn from).

```
class abc {
    public function do_something() { }
    
    public static function get_instance() {
        if ( null == self::$instance ) {
            self::$instance = new self;
        }
        return self::$instance;
    }
}

$abcSingleton = abc::get_instance();
$abcSingleton->do_something();
```

```
class def {
    public function do_something() { }
    
    public static function factory() {
        static $instance = false;
        if ( ! $instance ) {
            $instance = new self;
    }    
    
        return $instance;
    }
}

def::factory()->do_something();
```

The factory() and get_instance() functions seem equivalent - if instance doesn't exist, create it. Then return the single instance. 

And then when you retrieve the instance externally, the only difference seems to be that in the singleton example you assign the instance to a variable, and then call the method. Whereas you do it all in-line in the factory example.

What am I missing?